requirements/common.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Django==4.0.4
django-on-heroku==1.1.2
djangorestframework==3.13.1
idna==3.3
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pip==22.0.4
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
setuptools==58.1.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
